Revert "bugfixes and backwards compatibility (#534)"
This reverts commit 0a7a52b97608048989b1c93e1e708c2adc5998b8.
