+++ .gitignore
@@ -115,6 +115,3 @@ tests/proportion_test_diagnostics.csv
 
 # macOS
 *.DS_Store
-
-# Visual Studio Code project settings
-.vscode/
