--- docs/source/api_reference/exposure/data_transformations.rst
@@ -1,1 +1,0 @@
-.. automodule:: vivarium_public_health.exposure.data_transformations
